backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.13.1
cryptography==41.0.7
aiofiles==23.2.1
pypdf2==3.0.1
python-docx==1.1.0
pydub==0.25.1
httpx==0.26.0

